a tiny renaming / reorganizing
